+++ monad-raptor/Cargo.toml
@@ -11,7 +11,7 @@ bench = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bitvec = { workspace = true }
 rand = { workspace = true }
 rayon = { workspace = true }
-
 tracing = { workspace = true }
